% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_environments.r
\docType{data}
\name{config}
\alias{config}
\title{Flags/values to be used in the 'dashboards' scene}
\format{An object of class \code{environment} of length 13.}
\usage{
config
}
\description{
Flags/values to be used in the 'dashboards' scene
}
\keyword{datasets}
