% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_environments.r
\docType{data}
\name{connections}
\alias{connections}
\title{Environment to store db connections}
\format{An object of class \code{environment} of length 0.}
\usage{
connections
}
\description{
Environment to store db connections
}
\keyword{datasets}
