% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email.R
\name{e_emails}
\alias{e_emails}
\title{e_emails}
\usage{
e_emails(project, production_days = c(1:7))
}
\arguments{
\item{project}{a}

\item{production_days}{days of the week when production emails will be sent}
}
\description{
e_emails
}
