% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airquality.R
\name{get_airquality}
\alias{get_airquality}
\title{get_airquality
Gets the air quality, population weighted at county and national levels}
\usage{
get_airquality(impute_missing)
}
\arguments{
\item{impute_missing}{Do you want missing data imputed?}
}
\description{
get_airquality
Gets the air quality, population weighted at county and national levels
}
