% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_db.R
\name{get_db_connection}
\alias{get_db_connection}
\title{get_db_connection}
\usage{
get_db_connection(driver = config$db_config$driver,
  server = config$db_config$server, port = config$db_config$port,
  user = config$db_config$user, password = config$db_config$password)
}
\arguments{
\item{driver}{driver}

\item{server}{server}

\item{port}{port}

\item{user}{user}

\item{password}{password}
}
\description{
get_db_connection
}
