% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_weather.R
\name{get_weather}
\alias{get_weather}
\title{get_weather
Gets the weather, population weighted at county and national levels}
\usage{
get_weather()
}
\description{
get_weather
Gets the weather, population weighted at county and national levels
}
