% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{initialize}
\alias{initialize}
\title{initialize}
\usage{
initialize(package, package_dir = paste0("/dashboards/dashboards_",
  package))
}
\arguments{
\item{package}{a}

\item{package_dir}{a}
}
\description{
initialize
}
