% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_environments.r
\docType{data}
\name{logdata}
\alias{logdata}
\title{Environment to store logs
test}
\format{An object of class \code{environment} of length 1.}
\usage{
logdata
}
\description{
Environment to store logs
test
}
\keyword{datasets}
