% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email.R
\name{mailgun}
\alias{mailgun}
\title{mailgun}
\usage{
mailgun(subject, html = " ", to = NULL, bcc = NULL,
  attachments = NULL, inlines = NULL, include_footer = T,
  production_days = c(1:7), ...)
}
\arguments{
\item{subject}{a}

\item{html}{a}

\item{to}{a}

\item{bcc}{a}

\item{attachments}{a}

\item{inlines}{a}

\item{include_footer}{a}

\item{production_days}{Days that the production computer is planned to send an email}

\item{...}{a}
}
\description{
mailgun
}
