% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{msg}
\alias{msg}
\title{Enhanced Messaging}
\usage{
msg(txt, type = "msg", syscallsDepth = 2, newLine = FALSE,
  slack = FALSE)
}
\arguments{
\item{txt}{Text}

\item{type}{msg, warn, err}

\item{syscallsDepth}{The number of syscalls included in the message. Set to 0 to disable.}

\item{newLine}{Should there be a new line at the start of the message?}

\item{slack}{Should this also be posted to slack?}
}
\description{
Enhanced Messaging
}
