% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{path}
\alias{path}
\title{Dashboard folders}
\usage{
path(inside = "data_raw", ...)
}
\arguments{
\item{inside}{where it is inside}

\item{...}{an optional file}
}
\description{
This function finds folders according to the dashboard philosophy
}
