% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_weekly_action.R
\name{perform_action}
\alias{perform_action}
\title{perform_action}
\usage{
perform_action(file, value, dev_always_performs = FALSE)
}
\arguments{
\item{file}{File that stores value}

\item{value}{Value that will be checked against the saved file}

\item{dev_always_performs}{Does dev always perform action?}
}
\description{
perform_action
}
