% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_weekly_action.R
\name{perform_weekly_action}
\alias{perform_weekly_action}
\title{perform_weekly_action}
\usage{
perform_weekly_action(file, dev_always_performs = FALSE)
}
\arguments{
\item{file}{File that stores last used date}

\item{dev_always_performs}{Does dev always perform action?}
}
\description{
perform_weekly_action
}
