% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_db.R
\name{get_db_connection}
\alias{get_db_connection}
\title{get_db_connection}
\usage{
get_db_connection(driver = "MySQL", server = "db", port = 3306,
  user = "root", password = "example")
}
\arguments{
\item{driver}{driver}

\item{server}{server}

\item{port}{port}

\item{user}{user}

\item{password}{password}
}
\description{
get_db_connection
}
