% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email.R
\name{mailgun}
\alias{mailgun}
\title{mailgun}
\usage{
mailgun(subject, html = " ", to = NULL, bcc = NULL,
  include_footer = T, ...)
}
\arguments{
\item{subject}{a}

\item{html}{a}

\item{to}{a}

\item{bcc}{a}

\item{include_footer}{a}

\item{...}{a}
}
\description{
mailgun
}
