% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_rundate.R
\name{update_rundate}
\alias{update_rundate}
\title{update_rundate
Updates the rundate db tables}
\usage{
update_rundate(package, date_extraction, date_results, date_run)
}
\description{
update_rundate
Updates the rundate db tables
}
